requirements.txt
requests
httplib2
oauth2client
google-api-python-client
asyncio
cryptography
discord.py
dateparser
Babel
flask-cors
requests-cache
encrypted-mysqldb

pytest
pytest-asyncio
pytest-mock
coverage
hypothesis
